refactor: move helper utilities to separate module
